use folder for test_openai_hallucinated_tool_call_util
